+++ src/structures/Command.js
@@ -58,13 +58,6 @@ module.exports = {
   botPermissions: [],
   userPermissions: [],
   validations: [],
-  command: {
-    enabled: true,
-    aliases: [],
-    usage: '',
-    minArgsCount: 0,
-    subcommands: [],
-  },
   slashCommand: {
     enabled: true,
     ephemeral: false,